lidar/requirements.txt
numpy
matplotlib
tensorflow==2.5.0
waymo-open-dataset-tf-2-5-0
tqdm
